refactor(sessions): extract session-form logic into use-session-form-state hook
Move form state, TanStack Form setup, derived defaults application, and
store/game change handlers out of session-form.tsx into a new
use-session-form-state hook. Pure helpers (schema, buildDefaults, NONE_VALUE,
parseOptInt, numStrOrEmpty, types) are split to session-form-helpers.ts.

The component shrinks from 652 to ~360 lines and becomes a Props-driven
presentational view. Public interface of SessionForm is unchanged.
